Use MySQL for endpoint tests
Several adjustments:
- return UTC-timezone aware datetimes via GraphQL
- store datetimes without microseconds (precision of our database)
- add Size test data (otherwise FK insertion for box fails in test)
- update fixtures
