+++ back/boxtribute_server/graph_ql/scalars.py
@@ -1,4 +1,4 @@
-from datetime import datetime
+from datetime import datetime, timezone
 
 from ariadne import ScalarType
 
@@ -8,8 +8,7 @@ date_scalar = ScalarType("Date")
 
 @datetime_scalar.serializer
 def serialize_datetime(value):
-    if isinstance(value, str):
-        value = datetime.fromisoformat(value)
+    value = value.replace(tzinfo=timezone.utc)
     return value.isoformat()
