+++ back/test/data/size.py
@@ -1,0 +1,9 @@
+from boxtribute_server.models.size import Size
+
+
+def data():
+    return {"id": 1, "label": "small"}
+
+
+def create():
+    Size.create(**data())
