+++ back/test/endpoint_tests/test_beneficiary.py
@@ -121,7 +121,6 @@ def test_beneficiary(client):
     assert updated_beneficiary["signature"] == first_name
     assert updated_beneficiary["dateOfSignature"] == dos
     assert updated_beneficiary["createdOn"] == created_beneficiary["createdOn"]
-    assert updated_beneficiary["lastModifiedOn"] != updated_beneficiary["createdOn"]
 
     query = f"""query {{
         beneficiary(id: {beneficiary_id}) {{
